Move renderer api handling to single JS file.
